+++ test/e2e/repo/repo_management_test.go
@@ -1,4 +1,4 @@
-package e2e
+package repo
 
 import (
 	"context"
@@ -7,8 +7,11 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/argoproj/argo-cd/server/repository"
+	"github.com/argoproj/argo-cd/test/e2e/fixtures"
 	"github.com/argoproj/argo-cd/util"
 )
+
+var fixture = fixtures.NewFixture()
 
 func TestAddRemovePublicRepo(t *testing.T) {
 	repoUrl := "https://github.com/argoproj/argocd-example-apps.git"
