+++ test/e2e/testdata/hook/hook.yaml
@@ -1,13 +1,13 @@
 apiVersion: v1
 kind: Pod
 metadata:
-  name: hook
   annotations:
     argocd.argoproj.io/hook: Sync
+  name: hook
 spec:
   containers:
-    - name: main
-      image: alpine:latest
-      command:
+    - command:
         - "true"
+      image: "alpine:latest"
+      name: main
   restartPolicy: Never
